Moved fixture to App.js, app can render all data in fixture
